% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-contracts.R
\docType{class}
\name{ArchetypeClassifier-class}
\alias{ArchetypeClassifier-class}
\title{Virtual class for Stage 0.5b archetype classifiers}
\description{
Virtual class for Stage 0.5b archetype classifiers
}
