% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-contracts.R,
%   R/13b-component-interpretation.R
\name{association_applicable}
\alias{association_applicable}
\alias{association_applicable,CrossSectionalBinaryAssociationStrategy,StateTransitionData-method}
\alias{association_applicable,CrossSectionalContinuousAssociationStrategy,StateTransitionData-method}
\alias{association_applicable,CrossSectionalOrderedAssociationStrategy,StateTransitionData-method}
\title{Test whether an association strategy applies}
\usage{
association_applicable(strategy, data, values)

\S4method{association_applicable}{CrossSectionalBinaryAssociationStrategy,StateTransitionData}(strategy, data, values)

\S4method{association_applicable}{CrossSectionalContinuousAssociationStrategy,StateTransitionData}(strategy, data, values)

\S4method{association_applicable}{CrossSectionalOrderedAssociationStrategy,StateTransitionData}(strategy, data, values)
}
\arguments{
\item{strategy}{an \code{AssociationStrategy} implementation}

\item{data}{a \code{StateTransitionData} object}

\item{values}{aligned metadata values}
}
\value{
one logical value
}
\description{
Test whether an association strategy applies
}
