% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-contracts.R,
%   R/13b-component-interpretation.R,
%   R/13c-independent-time-course-interpretation.R,
%   R/13d-repeated-time-course-interpretation.R
\name{association_strategy_id}
\alias{association_strategy_id}
\alias{association_strategy_id,CrossSectionalBinaryAssociationStrategy-method}
\alias{association_strategy_id,CrossSectionalContinuousAssociationStrategy-method}
\alias{association_strategy_id,CrossSectionalOrderedAssociationStrategy-method}
\alias{association_strategy_id,IndependentTimeCourseLinearAssociationStrategy-method}
\alias{association_strategy_id,RepeatedTimeCourseLmerAssociationStrategy-method}
\title{Return an association strategy provenance identifier}
\usage{
association_strategy_id(strategy)

\S4method{association_strategy_id}{CrossSectionalBinaryAssociationStrategy}(strategy)

\S4method{association_strategy_id}{CrossSectionalContinuousAssociationStrategy}(strategy)

\S4method{association_strategy_id}{CrossSectionalOrderedAssociationStrategy}(strategy)

\S4method{association_strategy_id}{IndependentTimeCourseLinearAssociationStrategy}(strategy)

\S4method{association_strategy_id}{RepeatedTimeCourseLmerAssociationStrategy}(strategy)
}
\arguments{
\item{strategy}{an \code{AssociationStrategy} implementation}
}
\value{
one stable non-empty character identifier
}
\description{
Return an association strategy provenance identifier
}
