% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13b-component-interpretation.R
\name{atlas_exclusions}
\alias{atlas_exclusions}
\title{Extract metadata exclusions from an atlas}
\usage{
atlas_exclusions(atlas)
}
\arguments{
\item{atlas}{a \code{MetadataAssociationAtlas}}
}
\value{
a data frame with metadata fields and exclusion reasons
}
\description{
Extract metadata exclusions from an atlas
}
