% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{column_type_checker}
\alias{column_type_checker}
\title{Quick function for checking if a column is of the right type using data-masking}
\usage{
column_type_checker(data, column, type)
}
\arguments{
\item{data}{Data Frame or Tibble object}

\item{column}{Column you want to check}

\item{type}{\code{column}'s expected type}
}
\value{
a character vector
}
\description{
Quick function for checking if a column is of the right type using data-masking
}
\examples{
\dontrun{
check_text <- df \%>\% column_type_checker(text_var, "character")

if (check_text == "no") stop("Wrong type")
}
}
