% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversation_landscape.R
\name{conversation_landscape}
\alias{conversation_landscape}
\title{Shiny Application of a conversation landscape.}
\usage{
conversation_landscape(
  data,
  ...,
  id = document,
  text_var = text,
  colour_var,
  cleaned_text_var = clean_text,
  date_var = date,
  sentiment_var = sentiment,
  url_var = permalink,
  size = 2,
  x_var = V1,
  y_var = V2,
  type = "scattergl",
  colour_mapping = NULL
)
}
\arguments{
\item{data}{Data Frame}

\item{...}{Additional Columns to select}

\item{id}{Your ID column, if there is not one in your data, create one}

\item{text_var}{The original text variable for display in the data table}

\item{colour_var}{The variable you wish to map colour to (should be a string or factor)}

\item{cleaned_text_var}{The cleaned text variable for bigram and token plots}

\item{date_var}{Your date variable}

\item{sentiment_var}{Sentiment variable}

\item{url_var}{Variable which contains a url (preferably after using link_click)}

\item{size}{}

\item{x_var}{Variable which contains your x co-ordinates}

\item{y_var}{Variable which contains your y co-ordinates}

\item{type}{Type of the plotly output, unlikely to change from 'scattergl'}

\item{colour_mapping}{Which colours the colour variable should be mapped to}
}
\value{
A shiny application
}
\description{
Multiple tabs - one for cleaning and selecting, others for summary plots.
}
