% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_forms.R
\name{convertNL}
\alias{convertNL}
\title{Convert sitecode for NF and LB sites}
\usage{
convertNL(SiteCode)
}
\arguments{
\item{SiteCode}{Site code to convert.}
}
\value{
New sitecode.
}
\description{
Converts site codes for NF and LB to new NL code
}
\examples{
convertNL("NF001")
}
\keyword{internal}
