% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-contracts.R
\docType{class}
\name{Decomposer-class}
\alias{Decomposer-class}
\title{Virtual class for Stage 1 comparative decomposition strategies}
\description{
Virtual class for Stage 1 comparative decomposition strategies
}
