% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04a-decomposition-result.R
\name{DecompositionResult}
\alias{DecompositionResult}
\alias{dr_V_star}
\alias{dr_sigma}
\alias{dr_coords}
\alias{dr_warnings}
\alias{dr_V_k}
\alias{dr_sigma_k}
\alias{dr_coords_k}
\alias{dr_k}
\title{Construct a validated DecompositionResult}
\usage{
DecompositionResult(
  V_star,
  sigma,
  coords,
  warnings = character(),
  V_k,
  sigma_k,
  coords_k,
  k
)

dr_V_star(x)

dr_sigma(x)

dr_coords(x)

dr_warnings(x)

dr_V_k(x)

dr_sigma_k(x)

dr_coords_k(x)

dr_k(x)
}
\arguments{
\item{V_star}{numeric p-vector}

\item{sigma}{numeric K-vector}

\item{coords}{list of K numeric n-vectors (component 1)}

\item{warnings}{character vector}

\item{V_k}{p x k numeric matrix}

\item{sigma_k}{K x k numeric matrix}

\item{coords_k}{list of K numeric n x k matrices}

\item{k}{positive integer}
}
\value{
validated \code{DecompositionResult}
}
\description{
Construct a validated DecompositionResult
}
