% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/19-analysis-specification.R
\name{.analysis_spec_provenance}
\alias{.analysis_spec_provenance}
\title{Produce a serialisable list summary of an AnalysisSpecification for provenance}
\usage{
.analysis_spec_provenance(spec)
}
\arguments{
\item{spec}{an \code{AnalysisSpecification} object}
}
\value{
named list suitable for inclusion in a provenance record
}
\description{
Produce a serialisable list summary of an AnalysisSpecification for provenance
}
\keyword{internal}
