% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.labs_render}
\alias{.labs_render}
\title{Title}
\usage{
.labs_render(plot_type)
}
\arguments{
\item{plot_type}{Type of plot}
}
\value{
ggplot boiler code
}
\description{
Title
}
\examples{
\dontrun{
}
}
\keyword{internal}
