% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.plot_sentiment_distribution}
\alias{.plot_sentiment_distribution}
\title{Title}
\usage{
.plot_sentiment_distribution(df, sentiment_var = sentiment)
}
\arguments{
\item{df}{Data Frame or Tibble Object}

\item{sentiment_var}{Name of sentiment variable =}
}
\value{
ggplot object
}
\description{
Title
}
\examples{
\dontrun{

}
}
\keyword{internal}
