% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/18-sampling-design.R
\name{.sampling_design_provenance}
\alias{.sampling_design_provenance}
\title{Produce a serialisable list summary of a SamplingDesign for provenance}
\usage{
.sampling_design_provenance(design)
}
\arguments{
\item{design}{a \code{SamplingDesign} object}
}
\value{
named list suitable for inclusion in a provenance record
}
\description{
Produce a serialisable list summary of a SamplingDesign for provenance
}
\keyword{internal}
