% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{download_box}
\alias{download_box}
\title{Title}
\usage{
download_box(exportname, plot, width = 6, height = 4)
}
\arguments{
\item{exportname}{Name of export as a string}

\item{plot}{The plot to download}
}
\value{
A download handler
}
\description{
Title
}
\examples{
\dontrun{
}
}
\keyword{internal}
