% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-contracts.R
\docType{class}
\name{DynamicsEstimator-class}
\alias{DynamicsEstimator-class}
\title{Virtual class for Stage 2 dynamics estimation strategies}
\description{
Virtual class for Stage 2 dynamics estimation strategies
}
