% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_forms.R
\name{fill_forms}
\alias{fill_forms}
\title{Fill in HTML into nomination forms}
\usage{
fill_forms(file.path = NULL, rename = "")
}
\arguments{
\item{file.path}{String file path to the folder location where you want your data read in and saved. Default will be to your working directory.}

\item{rename}{Character string to append at end of original file name. If not provided the original file will be overwritten. Any files that already contain the renamed string will be ignored or overwritten. For example if "_HTML" is provided AB001.xlsm will be saved as a new file AB001_HTML.xlsm.}
}
\value{
New xlsm files with html text provied in the appropriate cells.
}
\description{
This function fills in the HTML nessisary for the final step in sending nomination forms for upload to KBA-EBAR.
}
\examples{
fill_forms(file.path = "D:/Key Biodiversity Areas/!Forms/FINAL", rename="_HTML" ))
}
\keyword{KBA,}
\keyword{forms,}
\keyword{html}
\keyword{nomination}
