% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-contracts.R
\docType{class}
\name{FitAssessor-class}
\alias{FitAssessor-class}
\title{Virtual class for Stage 0.75 distributional fit assessors}
\description{
Virtual class for Stage 0.75 distributional fit assessors
}
