% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-registry.R
\name{get_strategy}
\alias{get_strategy}
\title{Retrieve an implementation constructor from the registry}
\usage{
get_strategy(contract, name)
}
\arguments{
\item{contract}{character contract class name}

\item{name}{character implementation name}
}
\value{
constructor function
}
\description{
Retrieve an implementation constructor from the registry
}
