% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iba_protocol.R
\name{iba_protocol}
\alias{iba_protocol}
\title{Download IBA Protocol Data}
\usage{
iba_protocol(IBA = NULL, file.path = NULL, username = NULL)
}
\arguments{
\item{IBA}{String 5 letter alpha numeric site ID for the IBA in which you want to download data. This must be provided.}

\item{file.path}{String file path to the folder location where you want you data saved. Default will be to your working directory.}

\item{username}{String username for your NatureCounts account.}
}
\value{
A saved xlsx file with raw and totaled IBA Protocol data.
}
\description{
This function provides a useful way for querying and downloading Total and Raw IBA Protocol Data for a given site on one or multiple dates.
}
\examples{
iba_protocol(IBA = "AB015",file.path = "D:/Key Biodiversity Areas/!KBA_Data/IBAProtocol",username = "sample")
}
\keyword{IBA}
\keyword{NatureCounts}
\keyword{Protocol,}
\keyword{download,}
