% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_ibas_protocols.R
\name{list_ibas_protocols}
\alias{list_ibas_protocols}
\title{Get a list of sites with IBA Protocol Data}
\usage{
list_ibas_protocols(Province = NULL)
}
\arguments{
\item{Province}{String or character vector containing province codes for which you want to check for IBAs with IBA Protocol data.}
}
\value{
A dataframe of IBAs with IBA Protocol Data
}
\description{
This function provides a useful way for querying and downloading Total and Raw IBA Protocol Data for a given site on one or multiple dates.
}
\examples{
list_ibas_protocols(Procince=c("BC","ON"))
}
\keyword{IBA}
\keyword{NatureCounts}
\keyword{Protocol,}
\keyword{download,}
