% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{ls_create_note_area}
\alias{ls_create_note_area}
\title{Add a re-sizeable text area to take notes in!}
\usage{
ls_create_note_area()
}
\value{
a moveable shiny tect box
}
\description{
Add a re-sizeable text area to take notes in!
}
\examples{
\dontrun{
ui <- shiny::fluidPage(
ls_create_note_area()
)

}
}
