% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{ls_example}
\alias{ls_example}
\title{Tibble of examples for LandscapeR functions & tests}
\format{
A tibble with 10 rows and 5 columns
\describe{
\item{document}{Post ID}
\item{text}{Text variable for data table}
\item{permalink}{Link to original post}
\item{clean_text}{cleaner version of text variable for plots}
\item{date}{Date variable in ymd_hms format}
\item{setniment}{Sentiment variable}
\item{V1}{x co-ordinates}
\item{V2}{y co-ordinates}
}
}
\usage{
data("ls_example")
}
\description{
Tibble of examples for LandscapeR functions & tests
}
\keyword{internal}
