% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{ls_link_click}
\alias{ls_link_click}
\title{Prepare a URL column to be clickable in Shiny/Data Table}
\usage{
ls_link_click(df, url_var)
}
\arguments{
\item{df}{Data Frame or Tibble Object}

\item{url_var}{URL Column}
}
\value{
data frame with URL column edited to be clickable
}
\description{
Will allow you to click the hyperlink to load a URL, e.g. for selecting an image.
Make sure that DataTable is rendered with the argument 'escape = FALSE' or column will be all text.
}
\keyword{internal}
