% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_plot_group_facet.R
\name{ls_plot_group_facet}
\alias{ls_plot_group_facet}
\title{Create static plots for each level of a grouping variable.}
\usage{
ls_plot_group_facet(
  df,
  x_var = V1,
  y_var = V2,
  group_var,
  nrow,
  fill_colour = "blue",
  output = c("wrapped_plots", "list_of_plots", "wrapped_and_list")
)
}
\arguments{
\item{df}{Data Frame or Tibble object}

\item{x_var}{The variable containing x-coordinates}

\item{y_var}{The variable containing y-coordinates}

\item{group_var}{The grouping variable to iteratively highlight}

\item{nrow}{How many rows the all-plots-wrapped version should have}

\item{fill_colour}{The colour for the highlighted variable}

\item{output}{Whether to return one plot with all levels of grouping variable, separate plots or the wrapped version and the separate plots in a nested list.}
}
\value{
Dependent on input of \code{output}. See output documentation for more information.
}
\description{
For each level of the grouping variable you'll receive a plot with that level highlighted and every other post not highlighted.
}
\examples{
df <- ls_example
df \%>\% ls_plot_group_facet(group_var = cluster, fill_colour = "midnightblue", nrow = 3)

}
