% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{ls_plot_group_sent}
\alias{ls_plot_group_sent}
\title{Quickly plot group sentiment distributions as a perentage in LandscapeR shiny app}
\usage{
ls_plot_group_sent(
  df,
  group_var = cluster,
  sentiment_var = sentiment,
  type = c("percent", "volume"),
  title = "Grouped Sentiment Chart",
  bar_labels = c("none", "percent", "volume")
)
}
\arguments{
\item{df}{data frame}

\item{group_var}{Grouping variable, e.g. country, topic, cluster}

\item{sentiment_var}{Sentiment variable (categorical)}

\item{type}{Whether the plot should be of volume or py percentage. Accepts "percent" or "volume"}

\item{title}{The title of the plot, entered as a string.}

\item{bar_labels}{Whether to add the raw volume, percentage or neither to the bars}
}
\value{
Ggplot stacked bar chart with x and y coords flipped
}
\description{
Quickly plot group sentiment distributions as a perentage in LandscapeR shiny app
}
