% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_plot_group_static.R
\name{ls_plot_group_static}
\alias{ls_plot_group_static}
\title{Create a landscape plot and colour in each level of a grouping variable.}
\usage{
ls_plot_group_static(df, x_var = V1, y_var = V2, group_var, point_size = 0.1)
}
\arguments{
\item{df}{Data Frame or Tibble object}

\item{x_var}{The variable containing x-coordinates}

\item{y_var}{The variable containing y-coordinates}

\item{group_var}{The grouping variable to iteratively highlight}

\item{point_size}{How big each point should be}
}
\value{
a ggplot object
}
\description{
Default colour mapping ises the viridis palette, you can simply add different colour maps to the plot using standard ggplot syntax, as the return object is a ggplot object e.g. plot + \code{scale_colour_manual(values = ...)}
}
\examples{
df <- ls_example \%>\% dplyr::mutate(cluster = factor(cluster))
df \%>\% ls_plot_group_static(group_var = cluster)
}
