% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{ls_plot_group_vol_time}
\alias{ls_plot_group_vol_time}
\title{Quickly plot faceted volume of groups over time in a LandscapeR Shiny App}
\usage{
ls_plot_group_vol_time(
  df,
  group_var = group,
  date_var = date,
  unit = c("day", "week", "month", "quarter", "year"),
  nrow = 4
)
}
\arguments{
\item{df}{Data frame or tibble}

\item{group_var}{grouping variable e.g. country, cluster, topic etc.}

\item{date_var}{Variable which contains date information (can be datetime too I think)}

\item{unit}{A single unit of time fed into lubridate::floor_date  "week", "day", "month","quarter", "year"}

\item{nrow}{How many rows the plot should be shown in}
}
\value{
ggplot object of faceted bar charts
}
\description{
Quickly plot faceted volume of groups over time in a LandscapeR Shiny App
}
