% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_plot_static.R
\name{ls_plot_static}
\alias{ls_plot_static}
\title{Create a landscape plot without segmenting by groups}
\usage{
ls_plot_static(df, x_var = V1, y_var = V2, fill_colour = "black")
}
\arguments{
\item{df}{Data Frame or Tibble object}

\item{x_var}{The variable containing x-coordinates}

\item{y_var}{The variable containing y-coordinates}

\item{fill_colour}{What colour the points should be}
}
\value{
a ggplot obje t
}
\description{
Create a landscape plot without segmenting by groups
}
\examples{
df <- ls_example
df \%>\% ls_plot_group_static()
}
