% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{ls_plot_tokens_counter}
\alias{ls_plot_tokens_counter}
\title{Title}
\usage{
ls_plot_tokens_counter(
  df,
  text_var = .data$mention_content,
  top_n = 20,
  fill = "#0f50d2"
)
}
\arguments{
\item{df}{Data Frame or Tibble object}

\item{text_var}{Name of text variable}

\item{top_n}{Number of tokens to show}

\item{fill}{Colour - string or hexcode}
}
\value{
ggplot object
}
\description{
Title
}
\keyword{internal}
