% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{ls_plot_volume_over_time}
\alias{ls_plot_volume_over_time}
\title{Title}
\usage{
ls_plot_volume_over_time(df, .date_var, unit = "week", fill = "#0f50d2")
}
\arguments{
\item{df}{Data Frame or Tibble object}

\item{.date_var}{Name of date variable}

\item{unit}{Unit of time}

\item{fill}{Colour - string or hexcode}
}
\value{
a ggplot object
}
\description{
Title
}
\keyword{internal}
