% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_plotly_umap.R
\name{ls_plotly_umap}
\alias{ls_plotly_umap}
\title{Render a landscape plot in a shiny app}
\usage{
ls_plotly_umap(data, x, y, type, group_var, key, text_var, height)
}
\arguments{
\item{height}{}
}
\description{
Renders an interactive plotly plot with a
}
\section{Calling the function}{

Due to plotly's non-standard evaluation you'll need to feed in any column
names, such as \code{color}, \code{x}, \code{y}, \code{text}, \code{key}, as strings.
}

\examples{
{
LandscapeR::ls_example \%>\%
dplyr::mutate(cluster = factor(cluster)) \%>\%
 ls_plotly_umap(x = "V1", y = "V2", type = "scattergl",
 key = "document", text_var = "text", height = 600, color = "cluster")
 }
}
