% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_sentiment_over_time.R
\name{ls_sentiment_over_time}
\alias{ls_sentiment_over_time}
\title{View a sentiment and volume over time chart}
\usage{
ls_sentiment_over_time(
  df,
  sentiment_var = sentiment,
  date_var = date,
  unit = c("week", "day", "month", "quarter", "year")
)
}
\arguments{
\item{df}{Data Frame or Tibble object}

\item{sentiment_var}{Name of your sentiment variable}

\item{date_var}{Name of your date variable}

\item{unit}{Time unit to count sentiment by}
}
\value{
a ggplot object
}
\description{
View a sentiment and volume over time chart
}
