% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13b-component-interpretation.R
\name{plot.AssociationAbstention}
\alias{plot.AssociationAbstention}
\title{Plot an association abstention}
\usage{
\method{plot}{AssociationAbstention}(x, y, ...)
}
\arguments{
\item{x}{an \code{AssociationAbstention}}

\item{y}{ignored}

\item{...}{ignored}
}
\value{
a \code{ggplot} object
}
\description{
Plot an association abstention
}
