% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-ground-truth.R
\docType{class}
\name{PotentialGroundTruth-class}
\alias{PotentialGroundTruth-class}
\title{Ground truth for Stage 2 quasi-potential recovery tests}
\description{
Ground truth for Stage 2 quasi-potential recovery tests
}
