% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_site.R
\name{process_site}
\alias{process_site}
\title{KBA Bird Threshold Tests on Single Polygon}
\usage{
process_site(
  site = NULL,
  location = NULL,
  thresholds = NULL,
  file.name = NULL,
  username = NULL,
  timeout = 320
)
}
\arguments{
\item{site}{a simple features shape for a single site or polygon. Comes from read_sf.}

\item{location}{A string location for the site. This is needed for geographically confined populations or subspecies. Can either be the provincial code e.g. 'BC' or a single IBA code e.g. 'BC017.' If ignored these thresholds will be ignored.}

\item{thresholds}{data.frame of new or current thresholds. If missing will use default values.}

\item{file.name}{String of file name and path to save your excel file. Must be in .xlsx format.}

\item{username}{String username for your NatureCounts account.}

\item{timeout}{integer value for the number of seconds to wait for naturecounts query. Default is 320 seconds so only increase in case of failure.}
}
\value{
A excel spreadsheet with thresholds met for a given polygon. Returns nothing if no thresholds found.
}
\description{
This function provides a method for querying NatureCounts data for a given polygon and testing against KBA Thresholds.
}
\examples{
process_site(BCSite,location = "BC",thresholds = thresholdsfile, file.name = "BCSiteKBA.xlsx" ))
}
\keyword{Bird}
\keyword{KBA}
\keyword{NatureCounts,}
\keyword{Thresholds,}
\keyword{download,}
\keyword{polygon}
