% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{reactive_labels}
\alias{reactive_labels}
\title{Programmatically generate reactive labels from a prefix}
\usage{
reactive_labels(prefix, input)
}
\arguments{
\item{prefix}{type of plot e.g. 'sentiment'}

\item{input}{reading from Shiny's server \code{input} object - the list of all inputs.}
}
\value{
ggplot boiler-plate code with dynamically generated inputs
}
\description{
Programmatically generate reactive labels from a prefix
}
\examples{
\dontrun{
}
}
\keyword{internal}
