% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-container.R
\name{register_migration}
\alias{register_migration}
\title{Register a schema migration function}
\usage{
register_migration(from, to, fn)
}
\arguments{
\item{from}{character source schema version}

\item{to}{character target schema version}

\item{fn}{function(StateTransitionData) -> StateTransitionData}
}
\value{
invisible NULL
}
\description{
Register a schema migration function
}
