% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_forms.R
\name{specialCharacters}
\alias{specialCharacters}
\title{Remove html characters}
\usage{
specialCharacters(x)
}
\arguments{
\item{x}{Character string.}
}
\value{
Processed text.
}
\description{
Remove html characters
}
\keyword{internal}
