% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-ground-truth.R
\docType{class}
\name{SubspaceGroundTruth-class}
\alias{SubspaceGroundTruth-class}
\title{Ground truth for Stage 0/1 decomposition recovery tests}
\description{
Ground truth for Stage 0/1 decomposition recovery tests
}
