% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/techreviewsummary.R
\name{techreviewsummary}
\alias{techreviewsummary}
\title{Generate Technical Review Documents}
\usage{
techreviewsummary(file.path = getwd(), sitecodes = c())
}
\arguments{
\item{file.path}{String file path to the folder location where you want the technical reviews saved. Default will be to your working directory.}

\item{sitecodes}{Character string or vector of IBA sitecodes of tecnical reviews you want generated. Please use the original IBA code for sites in NF and LB.}
}
\value{
Word documents for each technical review done at sites provided.
}
\description{
This function takes generates word documents for technical reviews of specified sites
}
\examples{
fill_forms(file.path = "D:/Key Biodiversity Areas/!Forms/FINAL", rename="_HTML" ))
}
\keyword{KBA,}
\keyword{document}
\keyword{review,}
\keyword{technical}
\keyword{word}
