% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{titles_render}
\alias{titles_render}
\title{Title}
\usage{
titles_render(plot_type, input)
}
\arguments{
\item{plot_type}{Name of plot type as a string}
}
\value{
ggplot boilerplate code
}
\description{
Title
}
\examples{
\dontrun{
}
}
\keyword{internal}
