% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-ground-truth.R
\docType{class}
\name{TopologyGroundTruth-class}
\alias{TopologyGroundTruth-class}
\title{Ground truth for Stage 0.5 archetype topology tests}
\description{
Ground truth for Stage 0.5 archetype topology tests
}
