% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13a-visual-evidence-view.R
\name{visual_evidence_summaries}
\alias{visual_evidence_summaries}
\title{Extract visual-evidence summaries}
\usage{
visual_evidence_summaries(view)
}
\arguments{
\item{view}{a \code{VisualEvidenceView}}
}
\value{
a data frame
}
\description{
Extract visual-evidence summaries
}
