% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-contracts.R,
%   R/13b-component-interpretation.R,
%   R/13c-independent-time-course-interpretation.R
\name{associate_component}
\alias{associate_component}
\alias{associate_component,CrossSectionalBinaryAssociationStrategy,numeric-method}
\alias{associate_component,CrossSectionalContinuousAssociationStrategy,numeric-method}
\alias{associate_component,CrossSectionalOrderedAssociationStrategy,numeric-method}
\alias{associate_component,IndependentTimeCourseLinearAssociationStrategy,numeric-method}
\title{Compute one component association}
\usage{
associate_component(strategy, scores, values)

\S4method{associate_component}{CrossSectionalBinaryAssociationStrategy,numeric}(strategy, scores, values)

\S4method{associate_component}{CrossSectionalContinuousAssociationStrategy,numeric}(strategy, scores, values)

\S4method{associate_component}{CrossSectionalOrderedAssociationStrategy,numeric}(strategy, scores, values)

\S4method{associate_component}{IndependentTimeCourseLinearAssociationStrategy,numeric}(strategy, scores, values)
}
\arguments{
\item{strategy}{an applicable \code{AssociationStrategy} implementation}

\item{scores}{one numeric component-score vector}

\item{values}{aligned metadata values}
}
\value{
a normalized association result list
}
\description{
Compute one component association
}
