% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-contracts.R,
%   R/13b-component-interpretation.R
\name{association_strategy_id}
\alias{association_strategy_id}
\alias{association_strategy_id,CrossSectionalBinaryAssociationStrategy-method}
\title{Return an association strategy provenance identifier}
\usage{
association_strategy_id(strategy)

\S4method{association_strategy_id}{CrossSectionalBinaryAssociationStrategy}(strategy)
}
\arguments{
\item{strategy}{an \code{AssociationStrategy} implementation}
}
\value{
one stable non-empty character identifier
}
\description{
Return an association strategy provenance identifier
}
