% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscapeR-package.R
\docType{package}
\name{landscapeR-package}
\alias{landscapeR}
\alias{landscapeR-package}
\title{landscapeR: A landscape simulator for R.}
\description{
This package is aimed at simulating categorical landscapes on actual geographical realms,
starting from either empty landscapes or landscapes provided by the user (e.g. land use maps).
}
\details{
Check out the vignette illustrating the use of landscapeR.\cr
Also: \url{https://github.com/dariomasante/landscapeR}
}
\section{landscapeR functions}{

\itemize{
\item{\code{\link{makePatch}} creates a single patch in the landscape.}
\item{\code{\link{makeClass}} creates a group of patches belonging to the same class.}
\item{\code{\link{expandClass}} expands an existing class of patches.}
}
}
\author{
Dario Masante \email{dmasan@ceh.ac.uk}
}

