% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_plotly_umap.R
\name{ls_plotly_umap}
\alias{ls_plotly_umap}
\title{Render a landscape plot inside a shiny app}
\usage{
ls_plotly_umap(data, x, y, type, group_var, key, text_var, height, width)
}
\arguments{
\item{data}{Data Frame or tibble object, will likely need to be called
with brackets as it should be reactive}

\item{x}{X co-ordinate variable (as a string)}

\item{y}{Y co-ordinate variable (as a string)}

\item{type}{type of plotly chart to render, usually 'scattergl' for perforamcne}

\item{group_var}{Which variable to colouyr by (as a string)}

\item{key}{ID column (as a string)}

\item{text_var}{Text variable (as a string)}

\item{height}{Height of plot (as a strong)}

\item{width}{Width of plot (as a string)}
}
\value{
a plotly object
}
\description{
Renders an interactive plotly plot with a
}
\section{Calling the function}{

Due to plotly's non-standard evaluation you'll need to feed in any column
names, such as \code{color}, \code{x}, \code{y}, \code{text}, \code{key}, as strings.
}

\examples{
{
LandscapeR::ls_example \%>\%
dplyr::mutate(cluster = factor(cluster)) \%>\%
 ls_plotly_umap(x = "V1", y = "V2", type = "scattergl",
 key = "document", text_var = "text", group_var = "cluster",  height = 600, width = 800)
 }
}
