% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_site.R
\name{process_site}
\alias{process_site}
\title{KBA Bird Threshold Tests on Single Polygon}
\usage{
process_site(site = NULL, location = NULL, thresholds = NULL, file.name = NULL)
}
\arguments{
\item{site}{a simple features shape for a single site or polygon. Comes frome read_sf.}

\item{location}{A string location for the site. This is needed for geographically confinded populations or subspecies. Can either be the provincial code e.g. 'BC' or a single IBA code e.g. 'BC017.' If ignored these tresholds will be ignored.}

\item{thresholds}{data.frame of new or current thresholds. If missing will use defualt values.}

\item{file.name}{String of file name and path to save your excel file. Must be in .xlsx format.}
}
\value{
A excel spreadshead with thresholds met for a given polyon. Returns nothing if no thresholds found.
}
\description{
This function provides a method for querying NatureCounts data for a given polygon and testing against KBA Thresholds.
}
\examples{
process_site(BCSite,location = "BC",thresholds = thresholdsfile, file.name = "BCSiteKBA.xlsx" ))
}
\keyword{Bird}
\keyword{KBA}
\keyword{NatureCounts,}
\keyword{Thresholds,}
\keyword{download,}
\keyword{polygon}
